libply: Handle special case for BEGIN { exit(0); }
When exit() was called in the BEGIN, we expect it finishes the
session.  But actually it proceeds to the main loop and wait for more
action because it consumes the buffer for BEGIN probe.

Change the ply_load() to return struct ply_return so that it can pass
the exit status and skip the main loop.  This was found durint the
test in the next patch.